Make sure that the signed timestamp is different for each iteration
Reader may decide to not request a document if it already has cached
given Access Credential and the signed timestamp received in AUTH1
response is not newer than the cached one.

Set signed timestamp in each iteration to be 1 second past the previous
one, so that the Reader can make proper decision about requesting the
Access/Revocation Document.

Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>